==> notebooks/modeling/01-priceMovement/00-intro.ipynb <==
# Price movement

- Cil tohoto experimentu je zjednodusenne olabelovat data a pouzit klasicke ML modely, k nauceni modelu s vysokou precision.
- Labelovani je pomoci metriky maximalniho returnu za den a zaroven maximalniho poklesu, nez dojde k naplneni maxima.
- Body, ktere maji return vetsi nez 1% a zaroven neklesnou o vice nez 1% jsou olabelovany jako 1
