==> notebooks/modeling/01-priceMovement/02-price-movement-calculation.ipynb <==
%run ./../../utils/commonImports.py
%run ./../../utils/tradingImports.py
%matplotlib inline

# Load Data

db = 'D:\\Dropbox\\My work\\krypl-project\\sqlite\\ploniex-chart-data\\USDT_BTC.db'
data = load_trading_data(db, 'chart_data', from_date='2014-01-01', period='5min')

data.shape

# Experiment

Celý experiment spočívá v olabelovaní dat následujícím způsobem. Pokud daný bod při nákupu vydělá více než 1% tak True.
Snaha bude naučit klasifikátor tak, aby měl co nejlepší precision a zároveň provedl co nejvíce obchodů.

# Calculation

import time

def _max_row(day_df, col):
    i = day_df[col].idxmax()
    max_row = day_df.iloc[i, :]
    return max_row


def _min_row(day_df, col):
    i = day_df[col].idxmin()
    min_row = day_df.iloc[i, :]
    return min_row


def _max(day_df):
    close_max_row = _max_row(day_df, 'close')
    open_max_row = _max_row(day_df, 'open')
    return (close_max_row, 'close') if close_max_row['close'] > open_max_row['open'] else (open_max_row, 'open')


def _fulfilled_row(row, day_df, col, target):
    day_df['R'] = _return(row['close'], day_df[col])
    fulfilled = day_df.query(f"R >= {target}")
    if fulfilled.shape[0] > 0:
        return fulfilled.iloc[0, :]
    else:
        return day_df.iloc[-1, :]


def fulfilled(row, day_df, target):
    close_row = _fulfilled_row(row, day_df, 'close', target)
    open_row = _fulfilled_row(row, day_df, 'open', target)
    return (close_row, 'close') if close_row['close'] > open_row['open'] else (open_row, 'open')


def _return(v1, v2):
    return (v2 / v1) - 1

maxs = []
till_maxs = []
mins = []
df = data.copy()
DAY = 288
TARGET = 0.01

start_time = time.time()
for i in range(df.shape[0] - 2):
    if i % 1000 == 0:
        print(i, '%.3f s' % (time.time() - start_time))
        start_time = time.time()
    day_i = i + DAY if i + DAY < df.shape[0] else df.shape[0]
    row = df.iloc[i, :]
    day_df = reset_index_hard(df.iloc[i+1:day_i, :]).reset_index()
    
    fullfiled_row, fullfiled_col = fulfilled(row, day_df, TARGET)
    
    max_index = fullfiled_row['index'] if fullfiled_row['index'] > 0 else 1
    day_df = day_df.iloc[:max_index, :]
    min_row = _min_row(day_df, 'low')
    
    maxs.append(_return(row['close'], fullfiled_row[fullfiled_col]))
    till_maxs.append(fullfiled_row['index'])
    mins.append(_return(row['close'], min_row['low']))

maxs += [-1, -1]
till_maxs += [-1, -1]
mins += [-1, -1]

df['maxReturn'] = maxs
df['tillMax'] = till_maxs
df['minReturn'] = mins

df

write_tsv(df, 'D:\\Dropbox\\My work\\krypl-project\\analysis\\return-01-within-day.tsv')
