==> notebooks/modeling/01-priceMovement/03-price-movement-analysis.ipynb <==
%run ./../../utils/commonImports.py
%matplotlib inline

# Load Data

df_all = read_tsv('D:\\Dropbox\\My work\\krypl-project\\analysis\\return-01-within-day.tsv')
df = df_all[['date', 'maxReturn', 'tillMax', 'minReturn']].head(df_all.shape[0] - 2)

df.query('maxReturn >= 0.01').shape[0] / df.shape[0]

# Timestamp diff

data = df_all.copy()
diffs = data['timestamp'].shift(-1) - data['timestamp']
data['timestampDiff'] = diffs

diffs.value_counts()

zero_diff = data.query('timestampDiff == 1200')
zero_diff

# Basic Stats

query = '(maxReturn > 0.01)'
df_true = df.query(query)
df_false = df.query(f'not {query}')

df_true.describe()

df_false.describe()
