==> notebooks/modeling/01-priceMovement/04-make-labels.ipynb <==
%run ./../../utils/commonImports.py
%matplotlib inline

# Make Labels

output_path = 'D:\\Dropbox\\My work\\krypl-project\\dataLabeled\\return-01-loss-max-01-within-day.tsv'

df = read_tsv('D:\\Dropbox\\My work\\krypl-project\\analysis\\return-01-within-day.tsv')
df['label'] = (df['maxReturn'] > 0.01) & (df['minReturn'] > -0.01)

write_tsv(df, output_path)

df.head(10)

# Explore

df = read_tsv(output_path)
df.head()

trues = df.query('label')
falses = df.query('not label')

trues['price'] = trues['close']
falses['price'] = falses['close']

cols = ['timestamp', 'price']

write_tsv(trues[cols], '/tmp/Candles/trues.tsv')
write_tsv(falses[cols], '/tmp/Candles/falses.tsv')
