==> notebooks/modeling/01-priceMovement/06-test-train-validation-split.ipynb <==
%run ../../utils/commonImports.py
%matplotlib inline

# Load Data

data_split_root = 'D:\\Dropbox\\My work\\krypl-project\\dataSplit\\01-priceMovement'

featuresRoot = 'D:\\Dropbox\\My work\\krypl-project\\dataFeatures'
fileName = 'return-01-loss-max-01-within-day.tsv'
featuresFile = f'{featuresRoot}\\{fileName}'
featured = read_tsv(featuresFile).sort_values('timestamp')
featured = reset_index_hard(featured)

featured.head()

# Split for modeling and strategy testing

featured['date'].max()

Need to be run in jupyter hub

from utils.data import ModelingData

drop_cols = ['timestamp', 'period', 'open', 'high', 'low', 'close', 'volume',
       'quoteVolume', 'weightedAverage', 'date', 'maxReturn', 'tillMax',
       'minReturn', 'label']
modeling_data = ModelingData(featured, drop_cols, '2016-06-01', '2017-10-05')

modeling_data.write('D:\\Dropbox\\My work\\krypl-project\\dataSplit\\01-priceMovement')
