==> notebooks/modeling/01-priceMovement/10-conclusion.ipynb <==
# Conclusion

Zvolene basic featury nejsou evidentne moc descriptivni. Model se ucivelice nahodne -> ma 67% precision, coz tak cca odpovida procentu labelu v datech.
Pri pouziti modelu jako stretegii entry pointu dojde k velkemu prodelavani -> muzou za to velke mnozstvi FP, ktere prijdou pred velkym propadem.

### Napady dale

- Snazit se pochopit jednotlive featury na grafech
- Zamyslet se nad kvalitnejsim labelovanim
- Vyhlazeni casove rady (treba pomoci Savitzky Golay filteru)
